Safari-611.1.10/LayoutTests/media/modern-media-controls/controls-visibility-support/controls-visibility-support-controls-toggle-expected.txt

Testing the ControlsVisibilitySupport behavior when the controls attribute changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Media has loaded metadata
PASS media.videoTracks.length became 1
PASS mediaController.controls.visible is true

Turning controls off
PASS mediaController.controls.visible is false

Turning controls back on
PASS mediaController.controls.visible is true

PASS successfullyParsed is true

TEST COMPLETE
